backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
python-decouple>=3.6
drf-yasg>=1.21.0